refactor file name for sync v2
