+++ sync/evm.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"sync"
 
-	"github.com/Conflux-Chain/confura-data-cache/nearhead"
 	"github.com/Conflux-Chain/confura-data-cache/store"
 	"github.com/Conflux-Chain/go-conflux-util/blockchain/sync/evm"
 	"github.com/Conflux-Chain/go-conflux-util/blockchain/sync/poll"
@@ -14,24 +13,16 @@ import (
 	"github.com/sirupsen/logrus"
 )
 
-type CatchUpConfig struct {
-	Adapter evm.AdapterConfig
-	Poller  poll.CatchUpOption
-	Writer  store.BatchWriteOption
-}
-
 type Config struct {
-	CatchUp CatchUpConfig
+	CatchUp struct {
+		Adapter evm.AdapterConfig
+		Poller  poll.CatchUpOption
+		Writer  store.BatchWriteOption
+	}
 
 	Adapter evm.AdapterConfig
 	Poller  poll.Option
 	Writer  store.WriteOption
-}
-
-type NearHeadConfig struct {
-	Adapter evm.AdapterConfig
-	Poller  poll.Option
-	Writer  nearhead.WriteOption
 }
 
 type Worker struct {
@@ -108,28 +99,4 @@ func (worker *Worker) catchUp(ctx context.Context) {
 	wg.Wait()
 
 	logrus.WithField("next", poller.NextBlockNumber()).Info("Complete to sync data in catch up mode")
-}
-
-// StartNearHead starts to sync the near head data in a separate goroutine.
-func StartNearHead(ctx context.Context, wg *sync.WaitGroup, config NearHeadConfig, cache *nearhead.EthCache) error {
-	adapter, err := evm.NewAdapterWithConfig(config.Adapter)
-	if err != nil {
-		return errors.WithMessage(err, "Failed to create EVM adapter")
-	}
-
-	// sync from the finalized block that will never be reverted
-	finalizedBlockNumber, err := adapter.GetFinalizedBlockNumber(ctx)
-	if err != nil {
-		return errors.WithMessage(err, "Failed to retrieve finalized block number")
-	}
-
-	poller := poll.NewLatestPoller(adapter, finalizedBlockNumber, poll.ReorgWindowParams{}, config.Poller)
-	wg.Add(1)
-	go poller.Poll(ctx, wg)
-
-	writer := nearhead.NewWriter(cache, config.Writer)
-	wg.Add(1)
-	go process.Process(ctx, wg, poller.DataCh(), writer)
-
-	return nil
 }
